+++ docs/changes.rst
@@ -4,8 +4,8 @@ Changelog
 11.3.8 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- Docx report: make it possible to define extra text for the title per sector;
+  add a custom template for the French COVID tool
 
 11.3.7 (2020-06-18)
 -------------------